model_training/requirements.txt
econml
numpy
optuna
pandas
google-api-core
google-auth
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
PyYAML
scikit_learn
scikit_survival
xgboost
